app/people/serializers.py
from rest_framework import serializers
from .models import People

class PeopleSerializer(serializers.ModelSerializer):
    class Meta:
        model = People
        fields = (
            "name",
            "follower",
            "profile_url"
        )
