requirements.txt
Django
psycopg2
uWSGI
djangorestframework